+++ tests/support/helpers.py
@@ -53,8 +53,6 @@ import salt.utils.stringutils
 
 if salt.utils.platform.is_windows():
     import salt.utils.win_functions
-else:
-    import pwd
 
 log = logging.getLogger(__name__)
 
@@ -1622,15 +1620,6 @@ def win32_kill_process_tree(pid, sig=signal.SIGTERM, include_parent=True,
     gone, alive = psutil.wait_procs(children, timeout=timeout,
                                     callback=on_terminate)
     return (gone, alive)
-
-
-def this_user():
-    '''
-    Get the user associated with the current process.
-    '''
-    if salt.utils.platform.is_windows():
-        return salt.utils.win_functions.get_current_user(with_domain=False)
-    return pwd.getpwuid(os.getuid())[0]
 
 
 def dedent(text, linesep=os.linesep):
